+++ learnosity_sdk/utils/__init__.py
@@ -1,5 +1,5 @@
 from .lrnuuid import Uuid
 
 __all__ = [
-        "Uuid",
+        "Uuid"
         ]